requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
httpx==0.28.1
python-dotenv==1.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
tenacity==9.0.0
cachetools==5.5.1
pytest==8.3.4
ruff==0.8.4
